optimization

+++ PositronUnitySdk/PositronSdk/Assets/PositronSdk/Scripts/Client/GameEntitiesData/NetGameObject.cs
@@ -6,8 +6,8 @@ namespace Positron.Client.GameEntities
     [MessagePackObject]
     public struct NetGameObject
     {
-        [Key(0)] public ulong AssetIndex { get; set; }
-        [Key(1)] public ulong CreationId { get; set; }
+        [Key(0)] public ushort AssetIndex { get; set; }
+        [Key(1)] public ushort CreationId { get; set; }
         [Key(2)] public uint ObjectId { get; set; }
         [Key(3)] public uint OwnerClientId { get; set; }
         [Key(4)] public NetVector3 Position { get; set; }
